Organizing files into folders
These demo projects are getting big.  Organizing the stuff from the
copied project (render_particles_2D_GPU_p_on_p_collisions) into folders.
A number of these items will be deleted entirely as I approach particle
organization from an entirely different standpoint, but I'll deal with
that after it is organized.

The files are currently organized in filters in VS, but the folder
itself is a mess.  Now making folders that correspond to VS' filters
that I've set up.